LayoutTests/http/tests/referrer-policy/no-referrer/cross-origin-navigation-does-not-inherit-referrer-policy.html

<!DOCTYPE html>
<html>
<head>
<meta name="referrer" content="no-referrer">
<script src="/js-test-resources/js-test.js"></script>
</head>
<body>
<script>
description("Tests that no-referrer referrer policy does not persist across cross-origin PSON navigations.");
jsTestIsAsync = true;

const channelName = "referrer-policy-test";
const bc = new BroadcastChannel(channelName);
bc.onmessage = function(event) {
    referrer = event.data.referrer;
    // Referrer should be present because the no-referrer policy should NOT have persisted.
    shouldNotBe("referrer", "''");
    bc.close();
    finishJSTest();
};

// Open a noopener popup that will navigate cross-origin twice and then check the referrer.
window.open("/referrer-policy/no-referrer/resources/navigate-cross-origin.html?channel=" + encodeURIComponent(channelName), "_blank", "noopener");
</script>
</body>
</html>
