LayoutTests/http/tests/referrer-policy/no-referrer/cross-origin-navigation-does-not-inherit-referrer-policy-expected.txt

Tests that no-referrer referrer policy does not persist across cross-origin PSON navigations.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS referrer is not ''
PASS successfullyParsed is true

TEST COMPLETE
